Fix python setup and x86 support
Before this commit I had puppet using a virtualenv but that doesn't seem
to be working like I thought it would. I switched it to installing pip
packages to the system directly.

While testing this I also found that I needed to add a conditional for
when I am testing using the x86 ISO from the Raspberry Pi website.