Create a migration task and update config location
This commit includes a new task that will migrate an existing config
file for PiWeatherRock from the previously recommended install location
to the new one and ensure that it has the proper permissions on it.

This commit also changes the default location of the config file to
directly in the user's home directory. The end result of this is that
after applying this module there will be a new-style configuration
located at /home/pi/piweatherrock-config.json