Update pisetup task for when this already exists
Prior to this commit, if a user ran the pisetup task and had previously
installed this module on their Pi a step that was intended to get the
current code deployed to the Pi would actually just be skipped. The
changes here make it so that the latest version of the module is always
installed to the Pi when this task is run.